+++ src/renderer/component/viewOnWebButton/index.js
@@ -1,0 +1,3 @@
+import ViewOnWebButton from './view';
+
+export default ViewOnWebButton;